+++ src/auth/auth.service.ts
@@ -1,7 +1,9 @@
 import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
 import { RpcException } from '@nestjs/microservices';
+import * as bcrypt from 'bcrypt';
 import { PrismaClient } from 'generated/prisma';
-import { RegisterUserDto } from './dto';
+
+import { LoginUserDto, RegisterUserDto } from './dto';
 
 @Injectable()
 export class AuthService extends PrismaClient implements OnModuleInit {
@@ -21,9 +23,40 @@ export class AuthService extends PrismaClient implements OnModuleInit {
         throw new RpcException({ status: 400, message: 'User already exists' });
       }
       const newUser = await this.user.create({
-        data: { email, name, password },
+        data: {
+          email,
+          name,
+          password: bcrypt.hashSync(password, 10),
+        },
       });
-      return { user: newUser, token: 'ABC' };
+      const { password: __, ...rest } = newUser;
+      return { user: rest, token: 'ABC' };
+    } catch (error) {
+      throw new RpcException({ status: 400, message: error.message });
+    }
+  }
+
+  async loginUser(loginUserDto: LoginUserDto) {
+    const { email, password } = loginUserDto;
+    try {
+      const user = await this.user.findUnique({
+        where: { email: email },
+      });
+      if (!user) {
+        throw new RpcException({
+          status: 400,
+          message: 'User/Password not valid',
+        });
+      }
+      const isValidPassword = bcrypt.compareSync(password, user.password);
+      if (!isValidPassword) {
+        throw new RpcException({
+          status: 400,
+          message: 'User/Password not valid',
+        });
+      }
+      const { password: __, ...rest } = user;
+      return { user: rest, token: 'ABC' };
     } catch (error) {
       throw new RpcException({ status: 400, message: error.message });
     }