+++ changelog.md
@@ -18,3 +18,8 @@
 - npm i prisma --save-dev
 - npx prisma init # Esto crea un .env con DATA_URL el cual debemos configurar con la url de Mongo para crear el schema. Y en el archivo de prisma configurar provider = "mongodb" y crear el model "User"
 - npx prisma generate # Genera los scripts de prisma para usarlos en el proyecto (clases, interfaces, etc. a partir del modelo)
+
+## Video 137
+
+- npm i bcrypt
+- npm i --save-dev @types/bcrypt